use redirect plugin; update next.config
