+++ netlify/build-plugins/netlify-plugin-contextual-auth/index.js
