+++ netlify/build-plugins/netlify-plugin-contextual-auth/manifest.yml
