+++ netlify/build-plugins/netlify-plugin-storyblok-redirects/manifest.yml
@@ -1,0 +1,10 @@
+name: netlify-plugin-storyblok-redirects
+inputs:
+  - name: keyvar
+    description: The environment variable name to store the Storyblok API key
+    default: 'STORYBLOK_ACCESS_TOKEN'
+  - name: region
+    description: The Storyblok space region
+  - name: storypath
+    description: The path to the Storyblok stories (starts_with filter)
+    default: 'global-components/redirects'
